# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Tashi API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /qrcodes:
    get:
      operationId: listQRCodes
      tags: [qrcodes]
      summary: List all QR codes (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of QR codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QRCode"
    post:
      operationId: createQRCode
      tags: [qrcodes]
      summary: Create a QR code (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQRCodeRequest"
      responses:
        "201":
          description: QR code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QRCode"
  /qrcodes/scan:
    post:
      operationId: scanQRCode
      tags: [qrcodes]
      summary: Scan and validate a QR code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanQRCodeRequest"
      responses:
        "200":
          description: QR code scanned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanResult"
        "400":
          description: Invalid or already used QR code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    QRCode:
      type: object
      properties:
        id:
          type: integer
        qrNumber:
          type: string
        productId:
          type: integer
        productName:
          type: string
        points:
          type: integer
        status:
          type: string
          enum: [unused, used]
        createdAt:
          type: string
      required:
        - id
        - qrNumber
        - productId
        - productName
        - points
        - status
        - createdAt
    CreateQRCodeRequest:
      type: object
      properties:
        qrNumber:
          type: string
        productId:
          type: integer
      required:
        - qrNumber
        - productId
    ScanQRCodeRequest:
      type: object
      properties:
        qrNumber:
          type: string
      required:
        - qrNumber
    ScanResult:
      type: object
      properties:
        pointsEarned:
          type: integer
        totalPoints:
          type: integer
        productName:
          type: string
        message:
          type: string
      required:
        - pointsEarned
        - totalPoints
        - productName
        - message
...