# lib/api-zod/src/generated/types/createQRCodeRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tashi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQRCodeRequest {
  qrNumber: string;
  productId: number;
}