# lib/api-zod/src/generated/types/qRCode.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tashi API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QRCodeStatus } from "./qRCodeStatus";

export interface QRCode {
  id: number;
  qrNumber: string;
  productId: number;
  productName: string;
  points: number;
  status: QRCodeStatus;
  createdAt: string;
}