# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tashi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScanResult {
  pointsEarned: number;
  totalPoints: number;
  productName: string;
  message: string;
}